requirements.txt
streamlit
pandas
google-generativeai
python-dotenv # For loading environment variables from .env file